# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для записи к специалистам",
        "title": "LAPS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "94.247.129.222:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/specialists/{id}/work-experience": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет новую запись об опыте работы для специалиста",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опыт работы"
                ],
                "summary": "Добавить опыт работы специалисту по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные об опыте работы",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.WorkExperienceDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданной записи об опыте работы",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специалист не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
...
        "/work-experience": {
            "get": {
                "description": "Возвращает список опыта работы указанного специалиста",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опыт работы"
                ],
                "summary": "Получить список опыта работы специалиста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "specialist_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список опыта работы",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.WorkPlace"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специалист не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет новую запись об опыте работы для специалиста",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опыт работы"
                ],
                "summary": "Добавить опыт работы специалисту",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "specialist_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные об опыте работы",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.WorkExperienceDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданной записи об опыте работы",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специалист не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/work-experience/{id}": {
            "get": {
                "description": "Возвращает детальную информацию об опыте работы по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опыт работы"
                ],
                "summary": "Получить информацию об опыте работы по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опыта работы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные об опыте работы",
                        "schema": {
                            "$ref": "#/definitions/domain.WorkPlace"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Опыт работы не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию об опыте работы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опыт работы"
                ],
                "summary": "Обновить опыт работы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опыта работы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные об опыте работы",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.WorkExperienceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Опыт работы успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/rest.messageResponseType"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Опыт работы не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет запись об опыте работы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Опыт работы"
                ],
                "summary": "Удалить опыт работы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID опыта работы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Опыт работы успешно удален"
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Опыт работы не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.WorkExperienceDTO": {
            "type": "object",
            "required": [
                "company",
                "position",
                "start_year"
            ],
            "properties": {
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_year": {
                    "type": "integer"
                },
                "from_date": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "start_year": {
                    "type": "integer"
                },
                "to_date": {
                    "type": "string"
                }
            }
        },
        "domain.WorkPlace": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_year": {
                    "type": "integer"
                },
                "from_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "position": {
                    "type": "string"
                },
                "specialist_id": {
                    "type": "integer"
                },
                "start_year": {
                    "type": "integer"
                },
                "to_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "rest.errorResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "rest.messageResponseType": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}